re-arrange location of animation related scripts.
