+++ Assets/Scripts/Animation/AnimationEvents/AnimationEvent.cs.meta
