--- Assets/Scripts/Common/AnimationEvents.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6e71f611e62db624fa339392971a142a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
